backend/requirements.txt
# Django and API
Django==5.0.4
djangorestframework==3.15.2
django-cors-headers==4.3.1

# Async tasks
celery==5.5.3
redis==5.0.3

# Media processing
yt-dlp==2024.10.7
ffmpeg-python==0.2.0

# Database
psycopg2-binary==2.9.9

# Environment management
python-dotenv==1.0.1

# Utility libraries
requests==2.32.3

# (Optional for production / deployment)
gunicorn==22.0.0
whitenoise==6.7.0
